requirements.txt
fastapi==0.113.0
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
pydantic==2.5.1